rework distances to target provider -> provider group
Changes the underlying schema modeling for distance matrix so that
instead of having a table storing distances between two provider groups,
instead we have a matrix table that stores distances between a provider
and a provider group.

Adds a new "failure domain" distance type to the distance calculation
and loads that sets distances between node providers in the same rack to
"local", sets the distances from node providers to racks in the same row
in the same datacenter to "rack" and to different rows in the same
datacenter to "row" distance.