Correct mistake in constructing allocation items
We were incorrectly constructing the allocation item list by just
putting the first provider in the result for each resource class into an
AllocationItem, regardless of whether that provider's ID was in the
set-intersection with previous providers of other requested resources.

This patch corrects that behaviour and grabs the first matching provider
for all resource classes.