specify claim config in resource PoC tests
Adds the ability to specify a claim configuration (stored in a YAML doc)
for the runm-resource data model testing framework. There was already
the ability to specify a deployment configuration that described the
scale and profile of a deployment (racks, rows, sites, etc) and this
patch adds the equivalent functionality for specifying the details of a
claim request for resources.

Issue #2