partially solve full capability constraints
Adds capability constraint solving for any and forbid capabilities
within a CapabilityConstraint. Corrects a mistake where the list of
CapabilityConstraints that are contained in a ClaimRequestGroup were
being AND'd together instead of OR'd together.

Still need to modify the code to return an object describing the matches
and anti-matches for the constraint. This way, anti-matches (matches for
forbidden constraints) can be applied in a set subtraction against match
results from other constraints in the ClaimRequestGroup.

Issue #2
Issue #7