add MatchContext object for storing match state
During placement/scheduling, a series of constraints are processed.
Providers matching these constraints are either set-unioned or
set-intersected against the set of already-matched providers for
previous constraints.

This patch moves a bunch of code that was doing the following logic:

 matches = set()
 new_matches = constraint()
 if new_matches is not None:
   matches &= set(new_matches)
 else:
   matches = new_matches

to a new MatchContext object that is responsible for keeping the state
of matched providers as each of the constraints in the claim request is
processed, one by one.