pass excludes to resource constraint matching
If a capabilities or other constraint is processed before a resource
constraint, and that prior constraint identified providers that should
be excluded from the end result, we now store those excluded providers
in the ConstraintMatchResult.excludes attribute. This patch passes that
attribute on to the resource constraint query functions to ensure that
the resource-constraint-related query provides for the
`WHERE providers.id NOT IN ($EXCLUDE)` clause.