`runm property-schema get` and getSession() helper
This patch adds the CLI side of the property-schema get work. In doing
so, we modified the Session protobuffer message to have an actual
Partition object instead of a partition string. This will allow us to
use non-UUID identifiers for partitions in the future and cache the UUID
values found for a partition in the metadata service.

Modifies the PropertySchemaGetRerquest gRPC message payload to allow
specifying a partition, object type, and property key, along with an
optional version. All of these attributes are elements of the property
schema's key in backend storage (etcd).

Adds a getSession() utility method to the `runm` client where we can add
common code for constructing a Session protobuffer message from CLI
args, environv variables, and a configuration file.

Issue #23