increase speed of Docker builds
Pulls out building of a runm/base Docker image (which itself is just
golang:alpine with the pre dep-ensure'd vendor directory copied into the
GOPATH) into a separate Dockerfile in cmd/Dockerfile. The
component-specific Dockerfiles in cmd/runm/Dockerfile and
cmd/runm-metadata/Dockerfile now base from runm/base and do not execute
dep ensure in the Dockerfile. They only do `go build` to build the
single static binary that goes in the image.

This dramatically reduced the total build time, making iterative
development of runm components a whole lot nicer and quicker. The total
build time for all three images (runm/base, runm/runm, and
runm/metadata) is now around 20 seconds instead of the >2 minutes it was
previously taking. And just changing some files in a runmachine package
(and not changing any deps in vendor/) now results in builds of <10
seconds instead of the >2 minutes it was taking previously.

Issue #17