metadata: property-schema set
savepoint for work on `runm property-schema set`. Changes the
protobuffer model for PropertySchema to contain just the object type
code and partition UUID as a string instead of ObjectType and Partition
objects respectively. Also removes the version attribute of the property
schema since we're going to rely on etcd's generations to handle
concurrent writes.

Issue #21