add --bootstrap-token support to runm-metadata
Adds support for a new --bootstrap-token CLI option (and
RUN_METADATA_BOOTSTRAP_TOKEN environ var) that instructs the
`runm-metadata` server to ensure a one-time-use bootstrap token is
created in etcd.

If --bootstrap-token isn't specified or RUNM_METADATA_BOOTSTRAP_TOKEN is
unset (or set to ""), the `run-metadata` service to ensure that no
bootstrap token exists in etcd.

Issue #30