Use proto.Marshal() when writing object to etcd
I was mistakenly using proto.MarshalTextString() on the Partition
protobuffer message that gets stored at $ROOT/partitions/by-uuid/{UUID}.
Unfortunately, I thought MarshalTextString() was just a shortcut for
string(proto.Marshal()). I was wrong. It actually outputs the text-based
format of the protobuffer message and puts it in a string. What I wanted
was to serialize the protobuffer message to []byte and then pass that as
a string to etcd's KV.Put() method.

Fixes Issue #33