make abstract cursor interface PB-specific
Make the abstract cursor interface just Scan() in a protobuffer message.
This removes all the generic copyToDest() and cloneBytes() stuff that
was added to support copying in string keys. This just was never used.

Also adds a new SlicePBMessageCursor concrete type that will be used by
the ObjectList method coming in next patches.