standardize PartitionFilter to match object type
Changes the way that filters are constructed for partition get and list
to match the way we're filtering on object types. I put the new
PartitionFilter message type into the partition.proto file and am going
to be consistent in this type of filter specification, including adding
prefix ability to partition searches.