add scope to object type
We needed a way to indicate the scope in which object names are
guaranteed to be unique and whether or not a project and/or partition
needs to be associated with an object during creation. So, this patch
adds a new field to ObjectType called scope which is either PARTITION or
PROJECT.

Removes the `runm.partition` object type that was being added because
partition isn't an object type.