construct PartitionObjectFilters from ObjectFilter
builds a set of PartitionObjectFilter objects from an ObjectFilter. A
PartitionObjectFilter has a specific partition UUID and object type code
that is used when fetching object UUIDs from the indexes in etcd. An
ObjectFilter has a PartitionFilter and ObjectTypeFilter element that
each may be expanded into more than one partition or object type
(because of the prefix matching that can be specified with the
UsePrefix field in PartitionFilter and ObjectTypeFilter message
objects).

Issue #43