implement start of server-side object create
Adds the start of an ObjectSet implementation to the gRPC API for newly
created objects. The payload of the request is simplified: there is a
Before and an After image of the object to be created or changed. If
Before is nil, that means the user expects to be creating a new object
with the specified name. Therefore, we validate that the supplied After
image contains a valid partition, valid object type, etc, and then pass
the validated Object message to the backend storage layer where the
objects/by-uuid/ valued key is written along with the by-name index
entry pointing to the newly-created object's UUID.

Issue #43