change Object.PartitionUuid -> object.Partition
When doing `runm object create` in later patches, we construct an Object
protobuffer containing a partition and object type. However, for a
better end-user experience, we allow users to specify a partition name
instead of a UUID, and in the gRPC runm-metadata service, we translate
any partition names to the corresponding UUID. In order to have the
interface make more sense, this patch renames Object.PartitionUuid to
Object.Partition. This field will contain the name or UUID when being
sent from the client in the ObjectSetRequest gRPC API, and will only
contain the partition UUID when stored in etcd and whenever the storage
layer is involved.