non-object-specific list objects by filter
Adds a less efficient range scan on all objects with in-memory filtering
by partition, project and object type. When a PartitionObjectFilter is
passed to ObjectList that does not have a Search field, this method will
be used to grab Object messages matching the partition UUID, project and
object type code in the PartitionObjectFilter.

Issue #43