unknown partition or object type return 0 records
When doing `runm object list --filter "partition=unknown"` or `runm
object list --filter "type=unknown"`, we were not properly
short-circuiting and returning 0 records when there were no matches for
the partition or object type specified in the object filter.

This patch corrects that issue and checks that we found at least one
partition or object type when such a filter was received.

Issue #52
Issue #53