ensure newly-allocation message in cursor.Scan()
In order to avoid issues where a previously Scan'd protobuffer message
struct's values were Merge'd into the current cursor iteration's
protobuffer message scan target, this patch moves the definition of the
protobuffer message scan target inside the `for cur.Next()` loop,
thereby ensuring we get a newly-allocated message with no previously-set
message fields in it.

It would be nicer if protobuffer messages had some sort of Clear()
method that would blank out a previously-allocated message, but for now,
this is safe if not super-efficient.

Fixes Issue #56