correct issue in exec-runm-command with CLI args
I hate Bash. Seriously, it just needs to die in a fire along with Perl.

This patch fixes the scripts/exec-runm-command.sh script to use the
special "$@" notation in Bash to pass all the CLI arguments as-is that
were passed to scripts/exec-runm-command.sh to the `docker run` command.
Previously, I was stored "$@" in a Bash variable called EXEC_COMMAND,
but doing that fucks up the quoting on CLI args. For instance, if I ran:

./scripts/exec-runm-command object list --filter "partition=part0 project=proj0"

then doing EXEC_COMMAND="$@" would cause the "partition=part0
project=proj0" to unquote itself so the thing that would get passed
to `docker run` would be this:

object list --filter "partition=part0" project=proj0

Changing the ./scripts/exec-runm-command.sh to pass "$@" instead of
$EXEC_COMMAND corrected the issue. Oh, and passing "$EXEC_COMMAND"
doesn't work since then Bash encloses all of "object list --filter
\"partition=part0\" project=proj0" as a single quoted argument, like a
complete retard.

This is why Bash is a monstrosity of sizeable magnitude and needs to die
in a fire.

Fixes Issue #57