simplify objectsGetBySearch()
There was quite a bit of duplication in the objectsGetByObjectSearch()
and objectsGetBySearch() functions in the metadata storage layer. This
patch removes that duplication and flattens things into a single
objectsGetBySearch() function. This function handles efficient indexed
searches *first* and if not possible, drops into a less efficient
algorithm that grabs all objects and runs manual filters on all those
objects, building up a list of filtered objects using a sieve-like
pattern.