properly handle default object filter
Splits out the expansion of PartitionObjectFilter objects from a
protobuffer ObjectFilter object. Adds to this split-out object_filter.go
file a defaultObjectFilter() method of the metadata server that returns
a storage.PartitionObjectFilter that contains a Partition protobuffer
object representing the partition specified in the session. In this way,
we always guarantee that object listing always takes at least one
filter, and we also guarantee that if the user did *not* specify a
partition in their object filter, that we then filter by the session's
specified partition anyway.