add object get metadata server implementation
Adds the implement for object get operations. Re-uses most of the object
list server side implementation, checking to ensure when listing objects
with the supplied object get filter, that we only retrieve a single
object.

This modifies the ObjectGetRequest gRPC protobuffer message from a
simple string search to passing an ObjectFilter object. This was
necessary since object names are not unique unless within an object
type, project and/or partition scope.

Issue #43