metadata server gRPC API for object delete
Adds basic stub implementation for the ObjectDelete gRPC API method of
the metadata server. Changes the original sketched-out request and
response message format to align the way that objects are specified
(using an array of ObjectFilter messages just like ObjectList and
ObjectGet work) and remove the unnecessary Error message format and
replace with a simple array of string error messages. The error returned
from the ObjectDelete method indicates to look in the response's errors
collection for more details.

In the process, DRY's up some of the code that normalizes the slice of
ObjectFilter messages into a slice of storage.ObjectListFilter objects.

Issue #43