DRY up object delete and set's key string creation
The by-name index key for objects depends on whether the object's type
scope is project or partition-based. The ObjectDelete and ObjectSet
methods had duplicate code for determining this by-name index key. This
patch simply consolidates the code into a new objectByNameIndexKey()
method which does the work of fetching the object's type and
constructing the index key string appropriately.