remove use of abstract.Cursor for object storage
After using the abstract.Cursor interface for a month or more, I'm not
thrilled with it. Golang native slices are a better, more natural, way
of dealing with array operations, and so let's just use them. If I ever
decide to make a storage layer driver for metadata service over an
RDBMS, I'll wrap go/sql's Cursor interfaces behind the storage layer
itself.