begin introducing update path for prop-def set
Changes the `runm property-definition create` command to `runm
property-definition set`. We will be doing this for the `runm object
create` command as well. This makes the UX simpler: if you want to
create or update something, use the `runm {thing} set` command.

Fixes the JSONSchemaString() serialization of the apitypes
PropertySchema struct to output all the relevant fields.

Isse #21