move required to property defintion not schema
Moves the "required" field from the schema field to the property
definition object itself. This will be important when we generate the
JSONSchema for the object type. We can iteratively call
JSONSchemaString() for each property schema that will be stored for
properties on an object type and collectively grab all required property
keys by looking at the property definition's IsRequired field instead of
parsing the schema itself.

Modifies the table output for `runm property-definition list` to show
the Required? field but not the schema field itself.