fix object and prop-def get error return
There were two bugs in the property-definition and object get
implementations that were leading to an improper error about multiple
matching records being returned instead of the expected not found error.

The first issue was due to incorrect ObjectListFilter or
PropertyDefinitionFilter structs being produced from the
expandPropertyDefinitionFilter() method. This was due to
improperly-shadowed variables used to create the partitions and object
types lists used in creating the filter structs. The fix for this was to
declare the err, partitions and objTypes variables at the top of the
expandPropertyDefinitionFilter() method instead of using Golang's
automatic := variable instantiation. This is definitely not the first
time I've been bitten by Golang's awkward duplicity in how variables can
be instantiated (auto versus declared)...

The second issue was due to a silly logical error in the ObjectGet and
PropertyDefinitionGet gRPC API service method implementations that was
checking for len(objects) != 1 and always returning
ErrMultipleRecordsFound. Simply changing this to >1 and adding in a
check for len(objects) == 0 to return ErrNotFound fixed this particular
problem.

Fixes Issue #75