remove secondary indexing on property definitions
There was really no need to use secondary indexes for property
definitions. There's just not going to be many of them even for
relatively large deployments, so the secondary indexes under the
$PARTITION key namespaces were just bloat and complexity.

Also changes the way that filters are evaluated to consolidate the
matching code inside a PropertyDefinitionFilter.Matches() method.