begin standardizing Conditions and Matchers
Breaks out a new standardized interface-using method of matching stored
protobuffer messages against one or more conditions. The idea behind
this patch series is to have a generic way of constructing comparison
functors that may be passed to matcher methods that winnow results of
list operations.

The first type of protobuffer message we test this strategy out on is
the PropertyDefinition protobuffer message, which we now are storing in
a single primary key (property-definitions/by-uuid/) index.