reshape metadata partition interface
The metadata service's PartitionGet and PartitionList interfaces are
changed to have more specific PartitionFilter structures. While the API
service's PartitionFilter is a simple search string match, the metadata
service's PartitionFilter struct contains a UuidFilter and NameFilter
object which instruct the metadata service to look up a partition by
UUID or name. This is a pattern we will push to other objects stored in
the metadata service: the API service has a more user-friendly interface
while the metadata service has a more structured, verbose, but more
efficient interface for retrieving object information quickly and
precisely.