correct issue with metadata PartitionGet
The metadata service's PartitionGet gRPC API call was broken in a
previous commit due to a bad validation check on the modified Filter
payload. This patch fixes this and gets PartitionGet working again. Also
included in this patch is a rename of the API Server struct's
metaPartitionByUuid and metaPartitionByName methods to
partitionGetByUuid and partitionGetByName.