flesh out API service ProviderCreate
Adds an implementation of the API service's ProviderCreate gRPC API
method. Yes, we're returning to using a create method instead of a
single set method. The API service is responsible for unmarshaling the
YAML input, validating it, creating the object in the metadata service,
and calling the resource service to create the provider.

Issue #89