improve mysql container start scripts
When starting up a container with mysqld server, it's tricky to tell
when connectivity is actually achieved due to the way the mysql Docker
container is designed. The entrypoint script starts up mysqld *twice*:
once in a bootstrapping mode that creates the root mysql user, binding
on port 0, and a second time in "normal" operating mode, binding to port
3306.

This patch improves the mysql_start_container() Bash function to look in
the Docker logs for the *second* start of the mysqld server, which is
actually when incoming connections will be accepted.