DB storage layer for runm-resource
Adds the storage layer for the runm-resource service. The storage layer
for the runm-resource service is a traditional RDBMS. Initially, we're
just supporting MySQL, and this patch just adds the initial plumbing for
a Store class that wraps underlying database calls. The only thing the
code currently does is add a db_version table to the runm-resource
database in a Store.migrate() method that we'll use for schema and data
migrations on startup of the runm-resource service. Still lots of work
to do, though, including wiring up all the SQL for grabbing and saving
provider information, ensuring partition records exist, etc.

Issue #89