reshape ObjectTypeFilter in metadata service
This patch changes the structure of the ObjectTypeFilter from the more
API-focused simple string search to the more plumbing-layer focus of an
encapsulate CodeFilter. We will eventually be switching all filter
objects in the metadata service to the more structured approach here,
with the runm client passing the less-structured simple search string
filters and the API service transforming those as appropriate to the
more complex but efficient structures used in the plumbing layer.