standardize API protobuffer filter messages
Modifies the PartitionFilter and ProviderFilter messages in the API
service to use a generic SearchFilter message that has a simple search
string and an optional prefix boolean. These filter messages present a
simple and consistent interface to the `runm` client and end user. They
are transformed into the lower-level plumbing-layer UuidFilter or
NameFilter by the API service when calling the metadata service.