rework providersGetMatching to be API-focused
Moves the providersGetMatching function out of
pkg/api/server/resource_client.go and into the
pkg/api/server/provider.go file. ProviderList now is a simple wrapper
around a call to providersGetMatching that streams the resulting slice
of API Provider messages back to the caller. This will allow us to reuse
this code in the ProviderDelete API call paths.