first part of reworking property definitions
We begin the process of reworking the entire property definition
infrastructure in this patch. A property definition is being modified to
be an object definition for a partition and object type combination.
Object definitions will have zero or more property definitions that are
supplied to the API service describing constraints on the properties
that may be attached to a specific object type in a partition.

The metadata service is no longer going to be storing separate property
definition objects, one for each property that has had a property
definition created for it. Instead, the metadata service will only be
storing an ObjectDefinition object, and that ObjectDefinition will have
a single schema field that is the constructed JSONSchema document that
describes all of the attributes AND properties that may be associated
with that object type.

The next patch in this series begins to shape the API service's new
ProviderDefinitionSet API call to construct this JSONSchema and send it
over the wire to the metadata service.