handle property permissions with object definition
Validates and stores property permissions along with the schema for a
provider definition (which is generalized to an object definition when
saved to the metadata service).

We ensure that properties listed in the provider definition are
writeable and readable by at least the project creating the property
definition and that if no permissions were listed for the property, we
add a global read permission to the property by default.