slight refactoring of provider create validation
Moves check of partition existence into the
validateProviderCreateRequest method and changes the return value to a
pb.Provider object instead of an apitypes.Provider object. Also modifies
the providerCreate() method to inline modify the supplied pointer to a
pb.Provider object instead of returning another pointer to a pb.Provider
object.

I'm going to use this "inline modify a supplied pointer" technique for
all the xxxCreate() methods and simplify the calling structure to only
return an error object.