make objectCreate() to modify supplied Object
Simplify the call signature of the API server's objectCreate() method to
simply modify the supplied pointer to the Object to create instead of
passing back a pointer to a new Object message. This makes these methods
simply return an error if anything went wrong, which is a simpler
calling signature.