validate provider objects with object def schema
Pulls the JSONSchema from the object definition for runm.provider's in
the partition the new provider is being created in. Use this JSONSchema
to validate the supplied input.

Still need to do a bunch of cleanup and code commenting, but the basic
operation now works to prevent providers that don't meet the constraints
that are defined for providers.

Here's an example of creating a provider (west1-row1-rack1-node1) that
contains all the required properties and failing to create a provider
(east1-row1-rack1-node1) that I removed the "location.rack" property
from the YAML document file:

```
$ ./scripts/runm.sh provider definition set -f tests/data/definitions/runm.provider.yaml
ok
$ ./scripts/runm.sh provider create -f tests/data/objects/providers/west1-row1-rack1-node1.yaml
139ad3517046444fbb0a81c751cf0bfd
$ ./scripts/runm.sh provider create -f tests/data/objects/providers/east1-row1-rack1-node1.yaml
Error: Error: provider not valid:
- properties: location.rack is required
```