remove bootstrap command and server impl
Now that we have a `runm partition create` command, there's no reason
(at this point at least...) to have a `runm bootstrap` command. I'm sure
once we get going on the authn/z work, we'll run into some
chicken-and-egg problems that *may* need something like a bootstrap
command, but it will be added to the API service, not the metadata
service anyway. So, for now, just deleting all of the bootstrap code
entirely.