lay groundwork for provider-type specific defs
In order to implement definitions for providers that vary by provider
type, I needed to rework how object definitions were stored in the
metadata service as well as rework how object definitions were looked up
for different object types.

This patch changes the metadata service's ObjectDefinitionGet and
ObjectDefinitionSet gRPC API methods to be specific to
ProviderDefinitionGet and ProviderDefinitionSet. However, the concept of
an ObjectDefinition is actually made more generic by removing the
Partition attribute. This allows an ObjectDefinition to be defined for
any object type, any partition, any provider type (or consumer type, or
any other dimension or type) as well as be defined to apply as a global
default.

The `runm provider definition get` and `runm provider definition set`
CLI commands are both changed to allow a `--global` and `--partition`
CLI option. If the `--global` CLI option is used, the global default
definition for providers is set or returned. If the `--partition` CLI
option is used, the specified partition will have the definition for
providers overridden from the global default. If neither the `--global`
or `--partition` CLI option is specified, then the user's session
partition is used and will fall back to the global default if no
definition override is set for that partition.

Issue #99
Issue #103