rework object and provider definition docs
Fixes up the concepts.md documentation to better explain the differences
between object attributes and object properties. Changes the concepts
documentation to better define what an object and an object definition
is.

Adds user-guide documentation for how to view and set an override for a
provider definition in a partition.
Issue #109