use github.com/ghodss/yaml
Replaces the direct use of gopkg.in/yaml.v2 with the
github.com/ghodss/yaml package. This new package is useful for
correcting marshaling problems with the gopkg.in/yaml.v2 code and allows
us to simplify JSON and YAML parsing and serialization so that we can
eventually only decoarate the protobuffer files with `json:XXX`
annotations and will not need to use `yaml:XXX` annotations on custom
structs in pkg/api/types.

Issue #113