metadata PartitionGet split into UUID and name get
Splits the single PartitionGet gRPC method of the metadata server into
PartitionGetByUuid and PartitionGetByName gRPC methods. We are moving
towards a single protobuffer namespace and the metadata service's XXXGet
gRPC methods will no longer be open-ended with the ability to specify
*either* a UUID *or* a name. Instead, there will be two separate gRPC
methods for looking up objects by UUID or name. The XXXList methods will
continue to accept more free-form requests, however.

In order for this to work, we needed to modify the checkSession
stand-alone function in the metadata server package and convert it into
a method on the metadata Server struct. The new checkSession method now
converts the Session Partition field to a UUID if it is not already a
UUID, and returns an error if the Session's Partition is a name
referring to an unknown partition. When the API server is updated to
have its own checkSession method, we'll do the lookup/translation on the
API server side instead of the metadata service side.

Issue #111