metadata ObjectTypeGet -> ObjectTypeGetByCode
Since we're moving to a single protobuffer namespace, we need the
metadata service's gRPC methods to be more plumbing and less porcelain.
In the preceding patch, we changed the PartitionGet gRPC method of the
metadata service from one that accepts either a UUID or a name filter
and instead separated it into two methods, one accepting a UUID and
another accepting a name string.

This patch simply continues this pattern, making the ObjectTypeGet gRPC
method into ObjectTypeGetByCode.

Issue #111