Change ObjectDelete to ObjectDeleteByUuids
This patch modifies the ObjectDelete gRPC method in the metadata service
to be called ObjectDeleteByUuids. No other change was made to the
payload or response for this method.

The implementation of ObjectDeleteByUuids was updated to remove a TODO
that we had to create a single UuidsCondition that functioned like an IN
(<uuid>, <uuid>) expression. This UuidsCondition is introduced in this
patch and simplifies the code paths in the ObjectDeleteByUuids method
and reduces the number of Condition objects created and copied.

Issue #111