remove unused proto definitions from metadata
We should only add gRPC methods and protobuffer message types when we
actually need a method/type. There were a couple methods I thought would
be useful (and might be in the future) but are removed here to keep
things simple as we move towards a single protobuffer namespace.

Issue #111