metadata provider definition API make explicit
This patch keeps the more abstract provider definition get gRPC method
for the API service in tact but removes the abstraction from the
metadata gRPC methods that get provider definitions. We now have
separate gRPC methods on the metadata service for getting the global
default provider definition, for getting a provider type override
definition, a partition override definition, and an override for a
partition PLUS a provider type (the many different levels of overridding
that an admin can perform).

This makes the abstraction isolated to just the API service, which is
what we want, and makes explicit, clear and single-purpose methods in
the metadata service.

Issue #111