align API and resource service def for Provider
The end goal is to have a single namespace of protobuffer messages. To
get there, we need to align the slightly-different definitions of the
Provider message between the API and resource services.

This patch aligns the two by making both into a "full" message with
expanded relations/sub-objects for Partition and ProviderType.

Issue #111