merge protobuffer namespaces into one
Consolidates the protobuffer namespace into a single one (runm). In
order to do this, there were a couple necessary modifications.

For the metadata and resource services, the List gRPC methods were
renamed to Find. This isn't because different services cannot have the
same method names. This is because you cannot define two protobuffer
message types with the same name but different structures. For the API
service's List operations, the Filter messages are structured
differently (more abstract, less explicit than the resource and metadata
gRPC methods of the same name). So, in order to make things work, we now
have ProviderFindFilter and ProviderFilter, then former used for Find
methods in the resource service and the latter used for the List method
in the API service.

We still need some patches that clean up some of the duplicative copying
of properties but this patch does at least get rid of the three Session
message types that were identical.

Issue #111