properly instantiate Provider struct for Scan
When consolidating the Provider protobuffer representations for the API
server and resource server, I failed to properly instantiate the
Provider.Partition and Provider.ProviderType sub-struct fields when
Scan'ing from the SQL result. This was leading to a panic in the
resource service and an unknown error being returned to the API service.

Fixes Issue #121