show partition UUID and provider type code
In modifying the Provider protobuffer message type, we changed the
Provider.Partition field to a Partition protobuffer message type from a
string. Same for the Provider.ProviderType field.

We forgot, however, to change the code that prints a Provider object to
stdout in the runm client.

This patch fixes that oversight.