adds a reset.sh script
Before we get into a functional test framework, I wanted to make it
easier and quicker to tear down and repopulate the test environment in
Docker. Therefore, I've created a scripts/reset.sh script that kills the
etcd, metadata, api and resource services and blanks out the MySQL DB,
then restarts all those services and re-populates the DB tables (due to
the auto-schema-migration that occurs when runm-resource is started).

Also included is a separation of the different container startup
functionality into their own scripts/start-runm-<SERVICE>-container.sh
scripts, which should help prevent copy/paste errors in the script code
in the future.