error when provider delete matches no records
Previously, when executing `runm provider delete <nonexistingid>`, we
would return a misleading and confusing error message:

Error: at least one UUID is required.

While the above was technically correct (we were attempting to call the
resource service's ProviderDeleteByUuids method with an empty slice of
UUIDs), it was confusing.

With this patch, we now return ErrNoMatchingRecords when no provider
records were found.

Note that this patch does not change the existing behaviour of `runm
provider delete` that will return success (with NumDeleted value of 1)
when the user calls `runm provider delete` with an existing identifier
as well as a non-existing identifier.

Issue #118