move etcd connect/config to separate package
We're going to be moving to a different architecture which uses control
loops to handle events. These events will be triggered by etcd watches
and will mean that the runm-api service will no longer be communicating
with both the metadata service and the resource service. Instead, the
runm-api service will write the full object being created or updated to
the metadata service, and that will trigger watches to fire for any
other service that is listening for a particular key in etcd. That
service will then take some control action depending on whether the
object was created, updated, or deleted.

For example, a user may run `runm provider create`. This will send a
protobuffer message containing the provider information to the
`runm-api` service, which will validate the payload and ensure it meets
schema and authz constraints. After that, the `runm-api` service will
call to the runm-metadata service which will store the created Provider
object in etcd.

The `runm-resource` service will be watching a specific etcd key prefix
($ROOT/partitions/$PARTITION_UUID/objects/by-type/runm.provider/) and
will get notified of the new provider UUID. `runm-resource` will then be
able to perform the necessary actions, such as creating a record of the
provider in the relational datastore along with inventory, capabilities,
and group membership records after reading the object with that UUID
from the metadata service.

We have a ways to go before the above is a reality, but this first patch
simply moves the etcd configuration and connection routines out of the
metadata service and into a new pkg/etcdutil package where it can
eventually be used by multiple services other than the runm-metadata
service.