remove object relation expansion in Object
We are moving towards representing a Partition as just another Object.
In order to do that, we first need to do a bit of housekeeping in the
Object schema. This patch modifies the Object schema in the following
ways:

* Change object_type to object_type_code, partition to partition_uuid,
and project to project_external_id to indicate these fields are scalars
and not pointers to composed sub-objects
* Re-order the fields in Object to have the required fields
(object_type_code, uuid and name) as the first fields in the schema and
the optional fields (project_external_id and partition_uuid) numbered
afterwards