adds Generation field to Object
Adds an Object.Generation field to each stored object in the metadata
service. This is an incrementing integer field, set to 1 on initial
creation, that represents the number of times the object has been
updated over its lifetime. This Generation field will be used by other
services, notably the resource/placement service to ensure that it keeps
its own state aligned with the single source of truth in the metadata
service (as well as allow for watches to be set up).